Add randomized tests for scanInRange, fix bugs
